--- a/go/engine/cryptocurrency.go
+++ b/go/engine/cryptocurrency.go
@@ -109,6 +109,7 @@ func (e *CryptocurrencyEngine) Run(m libkb.MetaContext) (err error) {
 	}
 
 	sig, _, _, err := libkb.MakeSig(
+		m,
 		sigKey,
 		libkb.LinkTypeCryptocurrency,
 		sigInner,