--- a/go/engine/prove.go
+++ b/go/engine/prove.go
@@ -196,6 +196,7 @@ func (p *Prove) generateProof(m libkb.MetaContext) (err error) {
 	}
 
 	p.sig, p.sigID, _, err = libkb.MakeSig(
+		m,
 		p.signingKey,
 		libkb.LinkTypeWebServiceBinding,
 		p.sigInner,