--- a/go/engine/track_token.go
+++ b/go/engine/track_token.go
@@ -226,6 +226,7 @@ func (e *TrackToken) storeRemoteTrack(m libkb.MetaContext, pubKID keybase1.KID) 
 
 	sigVersion := libkb.SigVersion(*e.arg.Options.SigVersion)
 	sig, sigID, linkID, err := libkb.MakeSig(
+		m,
 		signingKey,
 		libkb.LinkTypeTrack,
 		e.trackStatementBytes,
@@ -264,7 +265,7 @@ func (e *TrackToken) storeRemoteTrack(m libkb.MetaContext, pubKID keybase1.KID) 
 		return err
 	}
 
-	me.SigChainBump(linkID, sigID)
+	me.SigChainBump(linkID, sigID, false)
 
 	return err
 }