… go/engine/revoke_test.go
… go/engine/user_test.go
… go/libkb/chain_link_v2_test.go
… go/libkb/sig_chain_test.go
… go/libkb/test_common.go
… go/libkb/user.go
… go/teams/create.go
… go/teams/rename.go
… go/teams/teams.go