### All writing of new elements in any sigchain links in this PR is behind a feature flag

#### For team sigchains
* add a `high_skip` node to the inner section of new team chain links that get sent to the server en route to other members. this uses the high skip values that https://github.com/keybase/client/pull/13387 set on the TeamSigchainState
* also hoist those two values up into the outer section of those team chain links
* add some test assertions that, after posting team chain links to the server, those chain links have the expected high_skip seqnos.

#### For user sigchains
* the same stuff. add high_skip to the inner and outer links. 
* hide it all behind a feature flag