@heronhaye 2018-08-28
WIP

todo:
- more tests
- many clarifications