@heronhaye 2018-08-29
I like the idea where clients can do whatever until a certain day when we say they must send hprev_info. Otherwise, if a user upgrades only on one device and posts an hprevinfo sig they'll be forced to upgrade their other devices too.

Sequence might be

1. Run backfills for teams and users
2. Merge in checking (optional) hprev infos on kbweb
3.1. Enable hprev support on go client 
3.2. Enable hprev support on webclient
4. (in the future) Enforce new sigs must be >2.3 on server