@patrickxb 2018-08-30
I wrote `client/go/tools/sigchain` to help benchmark sigchain loading.  It might be helpful to use it for this work.  It has support for memory and cpu profiling.