@heronhaye 2018-08-31
Note: I ended up going with the "forward loop after verify". If we change that logic it should be in another PR I think.

Here are some basic benchmarks with patrick's tool. I use Chris' sigchain up to link 1181--I was having difficulty inserting hPrevInfos into Sigchain v2 links and generating the sig properly.

Here's what's weird: My code runs faster than master (I've rebased) for even a sigchain without any hPrevInfos. I think this is probably due to a bug and I'm currently trying to find out why.

**New code, chain without any hPrevInfos**
sig chain load time: 220.659368ms
sig chain load time: 215.105665ms
sig chain load time: 185.585235ms
sig chain load time: 185.113498ms
sig chain load time: 181.354496ms
sig chain load time: 182.001488ms

**New code, chain specifies hPrevInfos at every link**
sig chain load time: 209.048189ms
sig chain load time: 187.376278ms
sig chain load time: 180.46583ms
sig chain load time: 183.717237ms
sig chain load time: 180.633384ms
sig chain load time: 180.596296ms


**Master, chain specifies hPrevInfos but is not parsed**
sig chain load time: 220.743641ms
sig chain load time: 199.590767ms
sig chain load time: 194.840192ms
sig chain load time: 200.595857ms
sig chain load time: 198.960194ms
sig chain load time: 196.902234ms