@xgess 2018-09-06
talking with @heronhaye, here's what's outstanding...

on this PR
1. stop actually sending `high_skip`s to the server. we can do this with the new feature flag implementation or by wrapping it in `if false` to stop adding complexity to an already very large piece of work, and then replace it with a feature flag in a fast-follow. 
2. start tracking high links on the team_sigchain_state as they come in. this was removed from an earlier commit and just needs to get put back.
3. get the server PR https://github.com/keybase/keybase/pull/2806 merged and deployed.
4. manual testing against pre-existing and new user/team chains on our deployed server code (with the feature flag off and on).


on the server PR (https://github.com/keybase/keybase/pull/2806) which validates any `high_skip`s that are on links...
1. maybe a little more manual testing on user sigchains, but otherwise it's ready for more eyes. kthx!


otherwise
1. finish backfill - might be worth discussing if it makes sense to go 100% of the way if we only *need* one pre-high_skip-high-link per sigchain. already brought this up with @mlsteele. 
2. set up feature flag on the server. off for everyone.
3. investigate how to build a reader/loader for teams that respects skips. we've been talking about what this will need to look like. what will have to be added to the server, how to fallback to the normal sigchain, etc. we'll make a design doc soon, but it likely won't be trivial. 
4. same for users. teams first because we think that will maybe be easier. i wouldn't be surprised if we change our minds an even number of times. 


think through roll out
1. are we going to get a new link type where an admin can sign for all the high links? @heronhaye mentioned it, but i was unfamiliar. maybe we can assume this will be there when we're building the reader so we can (1) take full advantage of it, and (2) reduce the scope of cases the reader will need to be able to handle. 
2. other rollout concerns like verifying that the clients and the server 100% agree on which are high links, do we want different feature flags for user chains and team chains, etc. all things we'll dig into. :)