@heronhaye 2018-09-10
There's still some more work to do before merging with making sure the upgrade process is smooth. I can change the variable name but I don't understand the issue--is it wrong or just confusing?