@heronhaye 2018-09-25
Miles' PR has a good list

    modern resets

    revocations

    web_service_binding that are eldest (danl 7)

    web_service_binding that are eldest on non-current subchains that are missing from the db