@xgess 2018-09-25
for convenience, here is that commit: https://github.com/keybase/client/commit/d75f95b9918a726698ea86877f86f2230b5fb8ac